requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
CodeforcesApiPy==2.0.11
Django==4.2.11
idna==3.6
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
